Removed duplicated data within Faction.

+++ HAPI_APP/src/Game.h
@@ -20,10 +20,10 @@ public:
 private:
 	std::array<Faction, static_cast<size_t>(FactionName::eTotal)> m_factions
 	{
-		static_cast<int>(FactionName::eYellow), 
-		static_cast<int>(FactionName::eBlue), 
-		static_cast<int>(FactionName::eGreen), 
-		static_cast<int>(FactionName::eRed)
+		FactionName::eYellow, 
+		FactionName::eBlue, 
+		FactionName::eGreen, 
+		FactionName::eRed
 	};
 	sf::RenderWindow m_window;
 	bool m_onlineGame;
